requirements.txt
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2
pydantic==2.10.2
tqdm==4.67.1
vaderSentiment==3.3.2
